Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,7 +6,7 @@ int main() {
    scanf("%d",&a);
    scanf("%d",&b);
    scanf("%d",&c);
-   int average = (a+b+c)/3;
-   printf("Average: %d",average);
+   float average = (a+b+c)/3.0;
+   printf("Average: %f",average);
     return 0;
 }
